Refactor: Article Service/Repository 합침
