+++ src/main/java/com/newzet/api/article/business/repository/ArticleRepository.java
@@ -1,9 +1,19 @@
 package com.newzet.api.article.business.repository;
 
 import java.util.List;
+import java.util.UUID;
 
 import com.newzet.api.article.business.dto.ArticleEntityDto;
+import com.newzet.api.article.repository.dto.ArticleWithImageProjection;
 
 public interface ArticleRepository {
 	List<ArticleEntityDto> saveAll(List<ArticleEntityDto> articleEntityDtoList);
+
+	List<ArticleWithImageProjection> getMonthlyArticleWithImage(UUID userId, int year, int month);
+
+	ArticleEntityDto getById(UUID articleId);
+
+	ArticleEntityDto readArticle(UUID articleId);
+
+	ArticleWithImageProjection getArticleByUserId(UUID userId);
 }
