+++ src/main/java/com/newzet/api/article/controller/ArticleController.java
@@ -9,12 +9,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.newzet.api.article.business.ArticleService;
+import com.newzet.api.article.business.service.ArticleService;
 import com.newzet.api.article.controller.dto.ArticleContentResponse;
 import com.newzet.api.article.controller.dto.ArticleListResponse;
-import com.newzet.api.auth.domain.AuthUser;
-import com.newzet.api.auth.infrastructure.annotation.Login;
-import com.newzet.api.auth.infrastructure.annotation.RequireAuth;
+import com.newzet.api.common.auth.annotation.Login;
+import com.newzet.api.common.auth.annotation.RequireAuth;
+import com.newzet.api.common.auth.domain.AuthUser;
 import com.newzet.api.common.response.ResponseCode;
 import com.newzet.api.common.response.SuccessResponse;
